+++ models/PropertyDoc.php
@@ -13,8 +13,8 @@ use yii\apidoc\helpers\PrettyPrinter;
 /**
  * Represents API documentation information for a `property`.
  *
- * @property bool $isReadOnly If property is read only. This property is read-only.
- * @property bool $isWriteOnly If property is write only. This property is read-only.
+ * @property boolean $isReadOnly If property is read only. This property is read-only.
+ * @property boolean $isWriteOnly If property is write only. This property is read-only.
  *
  * @author Carsten Brandt <mail@cebe.cc>
  * @since 2.0
@@ -34,7 +34,7 @@ class PropertyDoc extends BaseDoc
 
 
     /**
-     * @return bool if property is read only
+     * @return boolean if property is read only
      */
     public function getIsReadOnly()
     {
@@ -42,7 +42,7 @@ class PropertyDoc extends BaseDoc
     }
 
     /**
-     * @return bool if property is write only
+     * @return boolean if property is write only
      */
     public function getIsWriteOnly()
     {