+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Yii Framework 2 apidoc extension Change Log
 2.0.5 under development
 -----------------------
 
+- Enh #13: Allow templates to be specified by class name (tom--)
+- Enh #13: Added a JSON template to output the class structure as a JSON file (tom--)
+
 
 2.0.4 May 10, 2015
 ------------------