+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Yii Framework 2 apidoc extension Change Log
 2.0.5 under development
 -----------------------
 
+- Bug #25: Fixed encoding of HTML tags in method definition for params passed by reference (cebe)
 - Enh #13: Allow templates to be specified by class name (tom--)
 - Enh #13: Added a JSON template to output the class structure as a JSON file (tom--)
 - Enh: Added callback `afterMarkdownProcess()` to HTML Guide renderer (cebe)