+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Yii Framework 2 apidoc extension Change Log
 
 - Bug #25: Fixed encoding of HTML tags in method definition for params passed by reference (cebe)
 - Bug #10470: Fixed TOC links for headlines which include links (cebe)
+- Bug #37: Fixed error when extending Interfaces that are not in the current code base (cebe)
 - Enh #13: Allow templates to be specified by class name (tom--)
 - Enh #13: Added a JSON template to output the class structure as a JSON file (tom--)
 - Enh: Added callback `afterMarkdownProcess()` to HTML Guide renderer (cebe)